+++ tema2.json
@@ -12,7 +12,7 @@
             "music": {
                 "icon_path": "music_button",
                 "x": 1920,
-                "y": 696
+                "y": 888
             },
             "music_active": {
                 "icon_path": "music1",
@@ -99,12 +99,12 @@
                 "x": 1920,
                 "y": 30
             },
-            "config_down": {
+            "down": {
                 "icon_path": "pdown",
                 "x": 1920,
                 "y": 504
             },
-            "config_up": {
+            "up": {
                 "icon_path": "pup",
                 "x": 1920,
                 "y": 302
@@ -149,12 +149,46 @@
                 "x": 480,
                 "y": 320
             },
-            "askbox": {
-                "icon_path": "askbox",
-                "x": 500,
-                "y": 200
+            "music_main_window": {             
+                "window_x": 0,
+                "window_y": 0,
+                "window_sizex": 1920,
+                "window_sizey": 1080,
+                "window_icon_sizex": 180,
+                "window_icon_sizey": 180,
+            },
+            "radio_main_window": {             
+                "window_x": 0,
+                "window_y": 0,
+                "window_sizex": 1920,
+                "window_sizey": 1080,
+                "window_icon_sizex": 180,
+                "window_icon_sizey": 180,
+            },
+            "stream_main_window": {             
+                "window_x": 0,
+                "window_y": 0,
+                "window_sizex": 1920,
+                "window_sizey": 1080,
+                "window_icon_sizex": 180,
+                "window_icon_sizey": 180,
+            },
+            "spotify_main_window": {             
+                "window_x": 0,
+                "window_y": 0,
+                "window_sizex": 1920,
+                "window_sizey": 1080,
+                "window_icon_sizex": 180,
+                "window_icon_sizey": 180,
+            },
+            "tidal_main_window": {
+                "window_x": 0,
+                "window_y": 0,
+                "window_sizex": 1920,
+                "window_sizey": 1080,
+                "window_icon_sizex": 180,
+                "window_icon_sizey": 180,
             }
-
         }
     }
 }
